New pathfinding algorithm
I've added a BFS-style "flood-fill" that will run on the entire map each tick.
It produces a graph that allows me to very quickly get a path from/to anywhere
on the map.

The only downside is that I obviously cannot use it to find out the path an
enemy would take to something like a super-pellet, for which I still need A*.